fastapi-prabha/fastapi-back/requirements.txt
# annotated-types==0.6.0
# anyio==4.2.0
# certifi==2023.11.17
# charset-normalizer==3.3.2
# click==8.1.7
# exceptiongroup==1.2.0
# fastapi==0.109.0
# greenlet==3.0.3
# h11==0.14.0
# idna==3.6
# psycopg2-binary==2.9.9
# pydantic==2.5.3
# pydantic_core==2.14.6
# requests==2.31.0
# sniffio==1.3.0
# SQLAlchemy==2.0.25
# starlette==0.35.1
# typing==3.7.4.3
# typing_extensions==4.9.0
# urllib3==2.1.0
# uvicorn==0.26.0





annotated-types==0.6.0
anyio==4.2.0
bson==0.5.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
psycopg2-binary==2.9.9
py_eureka_client==0.11.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing==3.7.4.3
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0




